refactor: code review suggestions
